+++ tests/conftest.py
@@ -1,6 +1,9 @@
-import pytest
+import os
 import allure
+import pytest
 
+from typing import List
+from scripts.utils.work_with_data import get_network_list
 from func_timeout import func_timeout, FunctionTimedOut
 from scripts.utils.chain_model import Chain
 from scripts.utils.metadata_interaction import get_properties
@@ -32,6 +35,20 @@ def execute_with_timeout(timeout: int, function, args: tuple, error_message: str
             allure.attach(name="Error", body=error_message,
                           attachment_type=allure.attachment_type.TEXT)
         raise FixtureTimeoutError(error_message)
+
+
+def get_unique_subquery_urls(network_list: List[dict]) -> List[str]:
+    subquery_urls = set()
+    for chain in network_list:
+        external_api = chain.get('externalApi', {})
+        for api_type, api_list in external_api.items():
+            if isinstance(api_list, list):
+                for api in api_list:
+                    if api.get('type') == 'subquery':
+                        subquery_urls.add(api['url'])
+            elif isinstance(api_list, dict) and api_list.get('type') == 'subquery':
+                subquery_urls.add(api_list['url'])
+    return list(subquery_urls)
 
 
 @pytest.fixture(scope="module", params=collect_nodes_for_chains(get_substrate_chains()), ids=task_ids)
@@ -68,3 +85,10 @@ def chain_model(request) -> Chain:
         error_message=f'Timeout {FIXTURE_TIMEOUT} when initializing properties: {chain.name}'
     )
     return chain
+
+
+@pytest.fixture(scope="module")
+def subquery_projects():
+    network_file_path = os.getenv('CHAINS_JSON_PATH', "chains/v20/chains.json")
+    network_list = get_network_list('/' + network_file_path)
+    return get_unique_subquery_urls(network_list)
