+++ tests/test_subquery_is_synced.py
@@ -1,8 +1,5 @@
-import os
-import pytest
 import requests
 from typing import List
-from scripts.utils.work_with_data import get_network_list
 
 payload = "{\"query\":\"query{\\n  _metadata{\\n    chain\\n    lastProcessedHeight\\n    targetHeight\\n  }\\n}\",\"variables\":{}}"
 headers = {
@@ -11,27 +8,6 @@ headers = {
     'User-Agent': 'okhttp/4.11.0',
     'Content-Type': 'application/json; charset=UTF-8'
 }
-
-
-def get_unique_subquery_urls(network_list: List[dict]) -> List[str]:
-    subquery_urls = set()
-    for chain in network_list:
-        external_api = chain.get('externalApi', {})
-        for api_type, api_list in external_api.items():
-            if isinstance(api_list, list):
-                for api in api_list:
-                    if api.get('type') == 'subquery':
-                        subquery_urls.add(api['url'])
-            elif isinstance(api_list, dict) and api_list.get('type') == 'subquery':
-                subquery_urls.add(api_list['url'])
-    return list(subquery_urls)
-
-
-@pytest.fixture(scope="module")
-def subquery_projects():
-    network_file_path = os.getenv('CHAINS_JSON_PATH', "chains/v20/chains.json")
-    network_list = get_network_list('/' + network_file_path)
-    return get_unique_subquery_urls(network_list)
 
 
 def test_subquery_is_synced(subquery_projects):
